requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.2
future==1.0.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
inflect==7.4.0
jsonpath-ng==1.6.1
more-itertools==10.5.0
mysql-connector-python==9.0.0
ply==3.11
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.54
starlette==0.38.4
thrift==0.20.0
tqdm==4.66.5
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
